requirements.txt
# ─── Core Streaming ───────────────────────────────────────────────────────────
kafka-python==2.0.2
pyspark==3.4.1
delta-spark==2.4.0

# ─── Azure SDK ────────────────────────────────────────────────────────────────
azure-storage-file-datalake==12.14.0
azure-identity==1.14.0
azure-eventhub==5.11.4

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
pydantic==2.4.2
loguru==0.7.2
tenacity==8.2.3         # Retry logic

# ─── Dev / Testing ────────────────────────────────────────────────────────────
pytest==7.4.2
pytest-mock==3.11.1
faker==19.6.2
